录用费用,部分单位可以录用数量为小数

+++ jayud-oms/jayud-oms-model/src/main/java/com/jayud/oms/model/po/OrderPaymentCost.java
@@ -70,7 +70,7 @@ public class OrderPaymentCost extends Model<OrderPaymentCost> {
     private BigDecimal unitPrice;
 
     @ApiModelProperty(value = "数量")
-    private Integer number;
+    private BigDecimal number;
 
     @ApiModelProperty(value = "币种代码")
     private String currencyCode;
